helper/git: fix a bug when getting file_path from item.
sometimes, we need to use item[:indentifier]. Some times we have to
use item.identifier. It is so weird :))

the previous commits removed the git information in the page footer.
now we add it again. Any page with :virtual => true will be ignored
by the hook (aka they will not have a git information). this is ne
cessary for /tags/*/ -- they are dynamic/virtual